refactor(backend): split GameService helpers and match report template (phase 4)
Extract broadcast, audit/webhook, and rapid-pair logic into dedicated
modules, remove unused MatchFinishedError, and move the match report HTML
shell into app/match_report_template.py.
